feat: 完成 Rondo 生产级重构与 Blink 1.3.14 升级 升级 Blink Gradle 插件及 blink-common 至 1.3.14 重构账户、存储与跨服同步逻辑，保证资金操作原子性与数据一致性 将跨服模式统一为 MySQL 权威存储、Redis 缓存失效通知 增加有界经济快照缓存，优化 PlaceholderAPI 与 Vault 高频查询 修复 Vault 冷读返回默认余额的问题 新增获取全部已注册货币 ID 的公开 API 调整快照 API，区分纯内存探测与异步权威加载 使用系统时区计算兑换周期，移除 UTC 硬编码 加强金额边界、配置重载、日志恢复及异常处理 补充账户、存储、周期窗口、快照和 API 自动化测试 完善 CI、构建与发布流程 移除本地 VitePress 文档及 GitHub Pages 工作流

+++ src/main/kotlin/priv/seventeen/artist/rondo/currency/Currency.kt
@@ -1,3 +1,19 @@
+/*
+ * Copyright 2026 17Artist
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
 package priv.seventeen.artist.rondo.currency
 
 import java.math.BigDecimal
@@ -30,7 +46,7 @@ data class Currency(
     /** 是否可转账 */
     val transferable: Boolean,
     /** 转账税率 */
-    val transferTaxRate: Double,
+    val transferTaxRate: BigDecimal,
     /** 是否注册为 Vault 主货币 */
     val vaultPrimary: Boolean,
     /** 是否参与排行榜 */
